Update dependency version for d3code/pkg package to v0.0.0-20230529210300-25cc0c8e646d
This commit updates the version of the d3code/pkg dependency package to v0.0.0-20230529210300-25cc0c8e646d in both the go.mod and go.sum files. This update was necessary to ensure compatibility with the latest version of the package and to take advantage of any bug fixes or new features that have been added since the previous version.