Update version of github.com/d3code/pkg to v0.0.0-20230930071117-b4c11361435a
The previous version of github.com/d3code/pkg was v0.0.0-20230930052751-bb1990e3b554, but it has been updated to v0.0.0-20230930071117-b4c11361435a in order to incorporate the latest changes and improvements. This update ensures that the project is using the most recent and stable version of the package. The corresponding go.sum and go.mod files have also been updated accordingly.